Add database stuff

+++ src/main/java/com/github/maxopoly/MemeMana/model/MemeManaOwner.java
@@ -2,6 +2,6 @@ package com.github.maxopoly.MemeMana.model;
 
 public interface MemeManaOwner {
 
-	public int getID();
+	public long getID();
 
 }
